+++ RELEASENOTES.md
@@ -11,6 +11,7 @@
 *   Audio:
     *   Allow constant power upmixing/downmixing in DefaultAudioMixer.
     *   Add support for float PCM to `ChannelMappingAudioProcessor`.
+    *   Add support for float PCM to `TrimmingAudioProcessor`.
 *   Video:
     *   Add experimental `ExoPlayer` API to include the
         `MediaCodec.BUFFER_FLAG_DECODE_ONLY` flag when queuing decode-only input